Fixed char used as subscript error Changed int to  socklen_t for posix routines using sockaddr

+++ act.informative.c
@@ -893,7 +893,7 @@ void do_help (struct char_data *ch, char *argument, int cmd)
   if (!ch->desc)
     return;
 
-  for (; isspace (*argument); argument++);
+  for (; isspace ((int)*argument); argument++);
 
 
   if (*argument) {
