+++ interpreter.c
@@ -495,7 +495,7 @@ char *one_argument (char *argument, char *first_arg)
 
   do {
     /* Find first non blank */
-    for (; isspace (*(argument + begin)); begin++);
+    for (; isspace ((int)*(argument + begin)); begin++);
 
     /* Find length of first word */
     for (look_at = 0; *(argument + begin + look_at) > ' '; look_at++)
@@ -547,13 +547,13 @@ int is_abbrev (char *arg1, char *arg2)
 /* return first 'word' plus trailing substring of input string */
 void half_chop (char *string, char *arg1, char *arg2)
 {
-  for (; isspace (*string); string++);
+  for (; isspace ((int)*string); string++);
 
-  for (; !isspace (*arg1 = *string) && *string; string++, arg1++);
+  for (; !isspace ((int)(*arg1 = *string)) && *string; string++, arg1++);
 
   *arg1 = '\0';
 
-  for (; isspace (*string); string++);
+  for (; isspace ((int)*string); string++);
 
   for (; *arg2 = *string; string++, arg2++);
 }
@@ -860,10 +860,10 @@ int _parse_name (char *arg, char *name)
   int i;
 
   /* skip whitespaces */
-  for (; isspace (*arg); arg++);
+  for (; isspace ((int)*arg); arg++);
 
   for (i = 0; *name = *arg; arg++, i++, name++)
-    if ((*arg < 0) || !isalpha (*arg) || i > 15)
+    if ((*arg < 0) || !isalpha ((int)*arg) || i > 15)
       return (1);
 
   if (!i)
@@ -900,7 +900,7 @@ void nanny (struct descriptor_data *d, char *arg)
       d->character->desc = d;
     }
 
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
     if (!*arg)
       close_socket (d);
     else {
@@ -959,7 +959,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_NMECNF:             /* wait for conf. of new name */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
 
     if (*arg == 'y' || *arg == 'Y') {
       SEND_TO_Q ("New character.\n\r", d);
@@ -982,7 +982,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_PWDNRM:             /* get pwd for known player */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
     if (!*arg)
       close_socket (d);
     else {
@@ -1022,7 +1022,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_PWDGET:             /* get pwd for new player */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
 
     if (!*arg || strlen (arg) > 10) {
       SEND_TO_Q ("Illegal password.\n\r", d);
@@ -1040,7 +1040,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_PWDCNF:             /* get confirmation of new pwd  */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
 
     if (strncmp (crypt (arg, d->pwd), d->pwd, 10)) {
       SEND_TO_Q ("Passwords don't match.\n\r", d);
@@ -1055,7 +1055,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_QSEX:               /* query sex of new user  */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
     switch (*arg) {
     case 'm':
     case 'M':
@@ -1084,7 +1084,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_QCLASS:{
       /* skip whitespaces */
-      for (; isspace (*arg); arg++);
+      for (; isspace ((int)*arg); arg++);
       switch (*arg) {
       case 'm':
       case 'M':{
@@ -1157,7 +1157,7 @@ void nanny (struct descriptor_data *d, char *arg)
 
   case CON_SLCT:               /* get selection from main menu */
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
     switch (*arg) {
     case '0':
       close_socket (d);
@@ -1222,7 +1222,7 @@ void nanny (struct descriptor_data *d, char *arg)
     break;
   case CON_PWDNEW:
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
 
     if (!*arg || strlen (arg) > 10) {
       SEND_TO_Q ("Illegal password.\n\r", d);
@@ -1239,7 +1239,7 @@ void nanny (struct descriptor_data *d, char *arg)
     break;
   case CON_PWDNCNF:
     /* skip whitespaces */
-    for (; isspace (*arg); arg++);
+    for (; isspace ((int)*arg); arg++);
 
     if (strncmp (crypt (arg, d->pwd), d->pwd, 10)) {
       SEND_TO_Q ("Passwords don't match.\n\r", d);
