+++ modify.c
@@ -138,7 +138,7 @@ void quad_arg (char *arg, int *type, char *name, int *field, char *string)
     return;
 
   /* string */
-  for (; isspace (*arg); arg++);
+  for (; isspace ((int)*arg); arg++);
   for (; *string = *arg; arg++, string++);
 
   return;
@@ -347,7 +347,7 @@ char *one_word (char *argument, char *first_arg)
   found = begin = 0;
 
   do {
-    for (; isspace (*(argument + begin)); begin++);
+    for (; isspace ((int)*(argument + begin)); begin++);
 
     if (*(argument + begin) == '\"') {  /* is it a quote */
 
@@ -475,7 +475,7 @@ void show_string (struct descriptor_data *d, char *input)
       SEND_TO_Q (buffer, d);
 
       /* see if this is the end (or near the end) of the string */
-      for (chk = d->showstr_point; isspace (*chk); chk++);
+      for (chk = d->showstr_point; isspace ((int)*chk); chk++);
       if (!*chk) {
         if (d->showstr_head) {
           free (d->showstr_head);
